requirements.txt
streamlit
langchain
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
youtube-transcript-api
python-dotenv
sentence-transformers
